+++ packages/react-ui/src/widgets/FeatureSelectionWidgetUI/FeatureSelectionWidgetUI.js
@@ -40,6 +40,8 @@ const StylesWrapper = styled(Paper)(({ theme }) => ({
  * -->
  */
 function FeatureSelectionWidgetUI({
+  className,
+  sx,
   selectionModes,
   editModes = [],
   selectedMode,
@@ -72,7 +74,7 @@ function FeatureSelectionWidgetUI({
     : 'Click on the map to create a mask';
 
   return (
-    <StylesWrapper>
+    <StylesWrapper sx={sx} className={className}>
       <FeatureSelectionUIToggleButton
         icon={selectedModeData?.icon}
         hoverTooltip={hoverTooltip}
@@ -109,6 +111,7 @@ function FeatureSelectionWidgetUI({
 }
 
 FeatureSelectionWidgetUI.defaultProps = {
+  sx: undefined,
   enabled: false,
   tooltipPlacement: 'bottom',
   editModes: [],
@@ -122,6 +125,8 @@ const MODE_SHAPE = PropTypes.shape({
 });
 
 FeatureSelectionWidgetUI.propTypes = {
+  sx: PropTypes.any,
+  className: PropTypes.any,
   selectionModes: PropTypes.arrayOf(MODE_SHAPE.isRequired).isRequired,
   editModes: PropTypes.arrayOf(MODE_SHAPE.isRequired),
   selectedMode: PropTypes.string.isRequired,