+++ packages/react-widgets/src/widgets/GeocoderWidget.js
@@ -37,6 +37,8 @@ const SearchIcon = (args) => (
 /**
  * Renders a <GeocoderWidget /> component
  * @param  {object} props
+ * @param  {object} [props.className] - Material-UI withStyle class for styling
+ * @param  {object} [props.sx] - MUI5 for styling with sx prop
  * @param  {Function} [props.onError] - Function to handle error messages from the widget.
  */
 function GeocoderWidget(props = {}) {
@@ -69,7 +71,7 @@ function GeocoderWidget(props = {}) {
   }, [geocoderResult, dispatch]);
 
   return (
-    <StyledPaper elevation={2}>
+    <StyledPaper className={`${props.className}`} elevation={2} sx={props.sx}>
       {loading ? (
         <CircularProgress size={18} sx={svgStyle} />
       ) : (
@@ -93,6 +95,8 @@ function GeocoderWidget(props = {}) {
 }
 
 GeocoderWidget.propTypes = {
+  className: PropTypes.string,
+  sx: PropTypes.object,
   onError: PropTypes.func
 };